requirements.txt
Flask==2.3.2
PyJWT==2.8.0
flask-restful
flask-jwt-extended
pytest
requests
python-dotenv
cryptography
watchdog
flask-cors
